Centralize price parsing in IMarketApiClient. Removed parsing from MarketService
