+++ CryptocurrencyExchange/InfrastructureProject/Market/BinanceMarketApiClient.cs
@@ -1,7 +1,11 @@
-﻿namespace CryptocurrencyExchange.InfrastructureProject.Market
+﻿using Newtonsoft.Json;
+
+namespace CryptocurrencyExchange.InfrastructureProject.Market
 {
     public sealed class BinanceMarketApiClient : IMarketApiClient
     {
+        private const string USDT_SYMBOL = "USDT";
+
         private readonly HttpClient httpClient;
 
         public BinanceMarketApiClient(HttpClient httpClient)
@@ -9,13 +13,36 @@
             this.httpClient = httpClient;
         }
 
-        public async Task<string> GetPriceRawAsync(string symbol)
+        public async Task<decimal> GetUsdtPriceAsync(string symbol)
         {
-            using var response = await httpClient.GetAsync($"ticker/price?symbol={symbol}");
+            var normalizedSymbol = NormalizeSymbol(symbol);
+
+
+            using var response = await httpClient.GetAsync($"ticker/price?symbol={normalizedSymbol}");
             response.EnsureSuccessStatusCode();
 
-            return await response.Content.ReadAsStringAsync();
+            return ParsePrice(await response.Content.ReadAsStringAsync());
         }
+
+        private string NormalizeSymbol(string coinSymbol)
+        {
+            var normalizedSymbol = coinSymbol.ToUpper();
+            if (!normalizedSymbol.EndsWith(USDT_SYMBOL))
+                normalizedSymbol += USDT_SYMBOL;
+            return normalizedSymbol;
+        }
+
+        private decimal ParsePrice(string content)
+        {
+            var dto = JsonConvert.DeserializeObject<PriceResponseDto>(content)
+              ?? throw new InvalidOperationException("Empty price response");
+
+            if (dto.Price is null)
+                throw new InvalidOperationException("Price not found in response");
+
+            return (decimal)dto.Price;
+        }
+
     }
 
 }
